#416 Move creation of ConstructorCall to SequenceDiagramBuilder
When a user attempts to create a constructor call, they use the normal
"CallEdge" creation tool from the tool bar. This design decision was to
avoid forcing the user to use two very similar tools to create call and
constructor edges. However, internally if a constructor call is created,
a ConstructorEdge must be used, so it's necessary to convert the
original CallEdge to a Constructor Edge. This was initially done in the
DiagramCanvasController, leading to a special case and coupling with the
SequenceDiagram Builder. To avoid this design weakness, the conversion
from CallEdge to ConstructorEdge is moved to the SequenceDiagramBuilder.